refactor: category/products components
