+++ client-app/shared/catalog/composables/index.ts
@@ -1,5 +1,4 @@
 export * from "./useCategory";
 export * from "./useProduct";
 export * from "./useProducts";
-export * from "./useProductsFilters";
 export * from "./useRelatedProducts";
